Update README.md
Problem Description

Please make sure to read all parts of this document carefully.

Ahoy matey,

The year be 1700. A group of pirates recently acquired a computer from the British Navy (just go with it)

in a surprise raid on a British fort. Despite their affinity for tea, the British had recently installed Java on

the computer before it was plundered by the pirates. The pirates, being business savvy as they are, have

offered a contract worth one-thousand doubloons to you to help modernize their logistical systems. You

choose to ignore the ethical issues of piracy for those sweet, sweet doubloons and the opportunity to

apply your polymorphism, dynamic binding, and casting knowledge to a real-world problem.